Add colors, getPosition now getContent. logCanvas option added.
To add colors I had to wrap each character in a span tag. So for displaying the canvas in an HTML setting you can just use drawCanvas. But to properly display the canvas in the console you will need to use logCanvas (which strips out all span tags). getPosition was changed to getContent because I feel like getContent better describes the function (to retrieve a character at a certain position).