Many additions to macros
I first changed macro functions to macro actions. Added the ability to make changes to how the macro draws on a case by case bases (polymorphism). Added macro_changes for the macro type in macros, this way you can make the changes to macros inside other macros as well. Added ability to make an array of macro changes so it is now possible to have one set of changes be dynamic changes and the other static. Added the set_macro type which is currently not working (I may take it out since I implemented it before this commit was fully realized. But it may have a use, we will see.) . Documentation and a full publish to NPM coming soon.